[dx] update settings-store structure
