+++ src/shared/stores/settings/constants.ts
@@ -1,0 +1,6 @@
+
+export const THEME_MODES = {
+  LIGHT: "light",
+  DARK: "dark",
+};
+
