+++ src/shared/stores/settings/index.ts
@@ -1,0 +1,2 @@
+export {useSettingsStore} from './settings-store'
+export * from './constants'
